:sparkles: Preferences GUI
Using the Show Preferences command in the command prompt will open the preferences with General, Theme, and Extension tabs. The General tab allows changing the environment for launching command line programs and using trashcan options. The Theme tab allows the creation, editing, and changing themes. The Extensions tab allows for the creation, editing, and deleting of extensions. Have fun!